Separates themeChangingTest to a separate file.
